vision-service/pyproject.toml
[project]
name = "smartpick-vision"
version = "0.1.0"
requires-python = ">=3.12"
description = "FastAPI service for SmartPick AI computer vision and ergonomic scoring"
readme = "README.md"
authors = [{ name = "SmartPick AI" }]
dependencies = [
  "fastapi>=0.115",
  "uvicorn>=0.30",
  "pydantic>=2.7",
  "httpx>=0.27",
  "numpy>=2.0",
  "opencv-python-headless>=4.10",
  "onnxruntime>=1.17",
  "scipy>=1.13",
  "pillow>=10.0",
  "pyyaml>=6.0"
]

[project.optional-dependencies]
test = [
  "pytest>=8.0",
  "httpx>=0.27"
]

[tool.pytest.ini_options]
pythonpath = ["src"]
